# docs/static/api/ludus.openapi.yaml
openapi: 3.0.1
info:
  title: Ludus API
  description: >-
    This API controls Ludus user management, range deployment, power state, and
    testing state.
  version: 2.0.0
  license:
    name: GNU AGPLv3
    url: https://www.gnu.org/licenses/agpl-3.0.txt/
  contact:
    name: Ludus Discord
    url: https://discord.gg/HryzhdUSYT
...
paths:
...
  /license:
    get:
      summary: Retrieve the ludus license
      deprecated: false
      description: |
        Returns the ludus license
      operationId: getLicense
      tags:
        - Host Management
      parameters: []
      responses:
        '200':
          description: License Object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LicenseDataObject'
          headers: {}
      security:
        - JWT: []
        - ApiKeyAuth: []
  /diagnostics:
    get:
      summary: Diagnostics
      deprecated: false
      description: >-
        Get system information that can be helpful for debugging issues with
        Ludus.
      operationId: diagnostics
      tags: []
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  cpu:
                    type: object
                    properties:
                      cores:
                        type: integer
                      model:
                        type: string
                  storage_pools:
                    type: array
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                        type:
                          type: string
                        size:
                          type: integer
                        free:
                          type: integer
                        disk_type:
                          type: string
              example:
                cpu:
                  cores: 16
                  model: AMD Ryzen 7 8845HS w/ Radeon 780M Graphics
                storage_pools:
                  - name: ''
                    type: dir
                    size: 1966606663680
                    free: 376010477568
                    disk_type: unknown
          headers: {}
      security:
        - JWT: []
        - ApiKeyAuth: []
...
  /ansible/role/vars:
    post:
      summary: Get role variables for one or more Ansible roles
      deprecated: false
      description: >
        Returns the configurable variables (from defaults/main.yml) for one or
        more Ansible roles.

        These are the variables that can be set via role_vars in the range
        configuration.

        If a role is not found, an empty vars object is returned.
      operationId: GetRoleVars
      tags:
        - Ansible Management
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                roles:
                  type: array
                  items:
                    type: string
                  example:
                    - geerlingguy.docker
                    - lae.proxmox
              required:
                - roles
      responses:
        '200':
          description: Role variables retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  roles:
                    type: array
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                          example: geerlingguy.docker
                        global:
                          type: boolean
                          example: false
                        vars:
                          type: object
                          additionalProperties: true
                          properties: {}
              example:
                roles:
                  - name: geerlingguy.docker
                    global: false
                    vars:
                      docker_package: docker-ce
                      docker_package_state: present
                  - name: nonexistent.role
                    vars: {}
          headers: {}
        '400':
          description: Bad request - at least one role name is required
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: At least one role name is required
          headers: {}
      security:
        - JWT: []
        - ApiKeyAuth: []
  /ansible:
    get:
      summary: Retrieve available Ansible roles and collections
      deprecated: false
      description: |
        Returns available Ansible roles installed for this
        user on the ludus host
      operationId: getRolesAndCollections
      tags:
        - Ansible Management
      parameters:
        - name: userID
          in: query
          description: >-
            pass an optional userID to get roles and collections for a specific
            user
          required: false
          schema:
            $ref: '#/components/schemas/userID'
      responses:
        '200':
          description: Available roles and collections
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    name:
                      type: string
                      example: geerlingguy.java
                    version:
                      type: string
                      example: 2.3.2
                    type:
                      type: string
                      example: role
                    global:
                      type: boolean
                      example: false
          headers: {}
      security:
        - JWT: []
        - ApiKeyAuth: []
  /ansible/role/scope:
    patch:
      summary: Move or copy roles between global and local scopes
      deprecated: false
      description: >
        Move or copy one or more Ansible roles between global (all users) and
        local (current user) installation scopes.

        Only administrators can use this endpoint.
      operationId: MoveRoleScope
      tags:
        - Ansible Management
      parameters:
        - name: userID
          in: query
          description: The userID to impersonate for this request (Admins only)
          required: false
          example: JD
          schema:
            type: string
            pattern: ^[A-Za-z0-9]{1,20}$
      requestBody:
        content:
          application/json:
            schema:
              properties:
                roles:
                  type: array
                  items:
                    type: string
                  example:
                    - geerlingguy.java
                    - geerlingguy.docker
                global:
                  type: boolean
                  description: Target scope - true for global, false for local
                  example: true
                copy:
                  type: boolean
                  description: >-
                    If true, copy the role (keep source). If false, move the
                    role (remove source).
                  example: false
              required:
                - roles
                - global
              type: object
            examples: {}
      responses:
        '200':
          description: roles processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: array
                    items:
                      type: string
                    example:
                      - geerlingguy.java
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        role:
                          type: string
                        error:
                          type: string
          headers: {}
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: At least one role name is required
          headers: {}
        '403':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Only administrators can move roles between scopes
          headers: {}
        '500':
          description: Error moving/copying role
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Internal server error
          headers: {}
      security:
        - JWT: []
        - ApiKeyAuth: []
  /ansible/role:
    post:
      summary: Install or Remove an Ansible role
      deprecated: false
      description: |
        Install or Remove an ansible role by name or URL.
        Admins can pass a userID parameter to install a role for other users
        Action can be either "install or "remove".
      operationId: InstallRole
      tags:
        - Ansible Management
      parameters:
        - name: userID
          in: query
          description: >
            pass an optional userID to install for.

            Defaults to the user's range (admins can interact with other users'
            ranges)
          required: false
          schema:
            $ref: '#/components/schemas/userID'
      requestBody:
        content:
          application/json:
            schema:
              properties:
                role:
                  type: string
                  example: geerlingguy.java
                version:
                  type: string
                  example: 2.3.2
                force:
                  type: boolean
                  example: true
                action:
                  type: string
                  example: install
                global:
                  type: boolean
                  example: false
              type: object
            examples: {}
      responses:
        '201':
          description: role installed
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: string
                    example: 'Successfully installed: geerlingguy.java'
          headers: {}
        '403':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: You are not authorized to perform this ansible action
          headers: {}
        '500':
          description: Error installing role
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: 'Unable to install the ansible role: geerlingguy.java'
          headers: {}
      security:
        - JWT: []
        - ApiKeyAuth: []
  /ansible/role/fromtar:
    put:
      summary: Install an Ansible role from local directory
      deprecated: false
      description: |
        Install an ansible role from a local directory.
        Admins can pass a userID parameter to install a role for other users.
      operationId: InstallRoleTar
      tags:
        - Ansible Management
      parameters:
        - name: userID
          in: query
          description: >
            pass an optional userID to install for.

            Defaults to the user's range (admins can interact with other users'
            ranges)
          required: false
          schema:
            $ref: '#/components/schemas/userID'
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  example: ''
                force:
                  type: boolean
                  example: false
            examples: {}
      responses:
        '201':
          description: role installed
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: string
                    example: Successfully installed role
          headers: {}
        '400':
          description: Error installing role - bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Invalid boolean value
          headers: {}
        '403':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: You are not authorized to perform this ansible action
          headers: {}
        '500':
          description: Error installing role - server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: 'Unable to install the ansible role: <error>'
          headers: {}
      security:
        - JWT: []
        - ApiKeyAuth: []
...
  /user/{userID}:
    delete:
      summary: Remove a user from the system
      deprecated: false
      description: |
        Removes a user from the system. This action is only available to admins.
      operationId: deleteUser
      tags:
        - User Management
      parameters:
        - name: userID
          in: path
          description: The id of the users to delete
          required: true
          example: ''
          schema:
            $ref: '#/components/schemas/userID'
        - name: deleteDefaultRange
          in: query
          description: Flag to delete user's default range
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: user deleted
          headers: {}
        '400':
          description: userID not provided
          headers: {}
        '404':
          description: user not found
          headers: {}
      security:
        - JWT: []
        - ApiKeyAuth: []
...
  /user/all:
    get:
      summary: List all users
      deprecated: false
      description: |
        Get all users in Ludus. Available only to admins
      operationId: listAllUsers
      tags:
        - User Management
      parameters: []
      responses:
        '200':
          description: search results matching criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserObject'
          headers: {}
        '400':
          description: bad input parameter
          headers: {}
      security:
        - JWT: []
        - ApiKeyAuth: []
...
  /migrate/sqlite:
    post:
      summary: Migrate SQLite to PocketBase
      deprecated: false
      description: Migrates Ludus data from SQLite to PocketBase.
      operationId: migrateSQLiteToPostgreSQL
      tags:
        - Migration
      parameters: []
      responses:
        '200':
          description: Migration started
          headers: {}
        '500':
          description: Migration failed
          headers: {}
      security:
        - JWT: []
        - ApiKeyAuth: []
...
components:
  schemas:
...
    userID:
      type: string
      pattern: ^[A-Za-z0-9]{1,20}$
      description: |
        Must be a unique 1-20 character alpha-numeric string.
        Initials are commonly used.
      example: JD
...
    LicenseDataObject:
      type: object
      properties:
        entitlements:
          type: array
          items:
            type: string
            description: Individual entitlement code
          description: The entitlements granted by the license
          example: ENTERPRISE_PLUGIN
        licensed_to:
          type: string
          description: The name of the person or organization the license is licensed to
          example: John Doe
        expires_at:
          type: string
          format: date-time
          description: The date and time the license expires
          example: '2022-08-29T09:12:33.001Z'
        active:
          type: boolean
          description: Whether the license is active
          example: true
        message:
          type: string
          description: The message associated with the license
          example: 'License active, expires: 2022-08-29 14:12:33, licensed to John Doe'
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
    JWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
servers: []
security:
  - JWT: []
  - ApiKeyAuth: []